refact: use prettier and refactor
