+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "editor.rulers": [
+    80
+  ]
+}
